+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### vNEXT
 
 - Add `result()`, `currentResult()`, `variables`, `setOptions`, `setVariables` [PR #29](https://github.com/kamilkisiela/apollo-client-rxjs/pull/29)
+- Add generic types to `RxObservableQuery` [PR #30](https://github.com/kamilkisiela/apollo-client-rxjs/pull/30)
 
 ### v0.4.1